# artifacts/api-server/src/routes/admin.ts
import { Router, type IRouter } from "express";
import { requireAdmin, requireAdminStrict } from "../middlewares/supabaseAuth.js";
import { db, usersTable, ordersTable, productsTable } from "@workspace/db";
import { eq, desc } from "drizzle-orm";

const router: IRouter = Router();

function formatUser(u: any) {
  return { ...u, createdAt: u.createdAt?.toISOString?.() ?? u.createdAt };
}

router.get("/admin/users", requireAdmin, async (req, res) => {
  try {
    const users = await db.select().from(usersTable).orderBy(desc(usersTable.createdAt));
    return res.json(users.map(formatUser));
  } catch (err) { req.log.error(err); return res.status(500).json({ error: "Internal error" }); }
});

router.put("/admin/users/:userId/role", requireAdmin, async (req, res) => {
  try {
    const { role } = req.body;
    const userId = String(req.params.userId);
    const [user] = await db.update(usersTable)
      .set({ role, updatedAt: new Date() })
      .where(eq(usersTable.supabaseId, userId))
      .returning();
    if (!user) return res.status(404).json({ error: "Not found" });
    return res.json(formatUser(user));
  } catch (err) { req.log.error(err); return res.status(500).json({ error: "Internal error" }); }
});

router.post("/admin/users/:userId/ban", requireAdmin, async (req, res) => {
  try {
    const { isBanned } = req.body as { isBanned: boolean };
    const userId = String(req.params.userId);
    const [user] = await db.update(usersTable)
      .set({ isBanned, updatedAt: new Date() })
      .where(eq(usersTable.supabaseId, userId))
      .returning();
    if (!user) return res.status(404).json({ error: "Not found" });
    return res.json(formatUser(user));
  } catch (err) { req.log.error(err); return res.status(500).json({ error: "Internal error" }); }
});

router.delete("/admin/users/:userId", requireAdminStrict, async (req, res) => {
  try {
    const adminUser = (req as any).dbUser;
    const userId = String(req.params.userId);
    if (adminUser.supabaseId === userId) {
      return res.status(400).json({ error: "Cannot delete your own account" });
    }
    await db.delete(usersTable).where(eq(usersTable.supabaseId, userId));
    return res.json({ message: "User deleted" });
  } catch (err) { req.log.error(err); return res.status(500).json({ error: "Internal error" }); }
});

router.get("/admin/dashboard", requireAdmin, async (req, res) => {
  try {
    const [allOrders, allUsers, allProducts] = await Promise.all([
      db.select().from(ordersTable).orderBy(desc(ordersTable.createdAt)),
      db.select().from(usersTable),
      db.select().from(productsTable),
    ]);

    const paidOrders = allOrders.filter(o => o.status === "completed" || o.status === "verified");

    const totalRevenue = paidOrders.reduce((sum, o) => sum + parseFloat(o.totalAmount), 0);
    const soldValue    = totalRevenue;

    const totalCatalogValue = allProducts.reduce((sum, p) => sum + parseFloat(p.price), 0);
    const inStockValue      = allProducts.filter(p => p.inStock).reduce((sum, p) => sum + parseFloat(p.price), 0);

    const ordersByStatus = {
      pending:   allOrders.filter(o => o.status === "pending").length,
      verified:  allOrders.filter(o => o.status === "verified").length,
      completed: allOrders.filter(o => o.status === "completed").length,
      rejected:  allOrders.filter(o => o.status === "rejected").length,
    };

    const categoryMap: Record<string, { count: number; value: number }> = {};
    for (const p of allProducts) {
      if (!categoryMap[p.category]) categoryMap[p.category] = { count: 0, value: 0 };
      categoryMap[p.category].count++;
      categoryMap[p.category].value += parseFloat(p.price);
    }
    const productsByCategory = Object.entries(categoryMap).map(([category, data]) => ({
      category, ...data,
    }));

    const revenueByDay: Record<string, { revenue: number; orders: number }> = {};
    const now = new Date();
    for (let i = 29; i >= 0; i--) {
      const d = new Date(now);
      d.setDate(d.getDate() - i);
      const key = d.toISOString().split("T")[0];
      revenueByDay[key] = { revenue: 0, orders: 0 };
    }
    for (const o of paidOrders) {
      const key = o.createdAt.toISOString().split("T")[0];
      if (revenueByDay[key]) {
        revenueByDay[key].revenue += parseFloat(o.totalAmount);
        revenueByDay[key].orders++;
      }
    }

    const productMap = new Map(allProducts.map(p => [p.id, p]));
    const recentOrders = allOrders.slice(0, 8).map((order) => {
      const product = productMap.get(order.productId);
      return {
        ...order,
        totalAmount: parseFloat(order.totalAmount),
        discountAmount: parseFloat(order.discountAmount),
        gameDetails: order.gameDetails || {},
        product: product ? {
          ...product, price: parseFloat(product.price),
          tags: Array.isArray(product.tags) ? product.tags : [],
          dynamicFields: Array.isArray(product.dynamicFields) ? product.dynamicFields : [],
          createdAt: product.createdAt?.toISOString?.(),
        } : null,
        createdAt: order.createdAt?.toISOString?.(),
        updatedAt: order.updatedAt?.toISOString?.(),
      };
    });

    return res.json({
      totalRevenue,
      soldValue,
      totalCatalogValue,
      inStockValue,
      totalOrders: allOrders.length,
      pendingOrders: ordersByStatus.pending,
      totalUsers: allUsers.length,
      totalProducts: allProducts.length,
      ordersByStatus,
      productsByCategory,
      recentOrders,
      revenueByDay: Object.entries(revenueByDay).map(([date, data]) => ({ date, ...data })),
    });
  } catch (err) { req.log.error(err); return res.status(500).json({ error: "Internal error" }); }
});

export default router;