# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gaming Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active banners
      responses:
        "200":
          description: List of banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "201":
          description: Created banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    put:
      operationId: updateBanner
      tags: [banners]
      summary: Update banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "200":
          description: Updated banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Chat ──────────────────────────────────────────────────────────────────
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat messages for a session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageBody"
      responses:
        "201":
          description: Message sent with bot/staff reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
  /chat/sessions/{sessionId}:
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session and all its messages (admin)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Admin / Users ─────────────────────────────────────────────────────────
...
  /admin/users/{clerkId}/role:
    put:
      operationId: updateUserRole
      tags: [admin]
      summary: Update user role (admin)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /admin/users/{clerkId}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban or unban a user (admin)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /admin/users/{clerkId}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user (admin)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        linkUrl:
          type: string
          nullable: true
        active:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required: [id, title, active, sortOrder, createdAt]
    CreateBannerBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        linkUrl:
          type: string
          nullable: true
        active:
          type: boolean
        sortOrder:
          type: integer
      required: [title, active, sortOrder]
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        sender:
          type: string
          enum: [user, bot, staff]
        content:
          type: string
        createdAt:
          type: string
      required: [id, sessionId, sender, content, createdAt]
...
    SendChatMessageBody:
      type: object
      properties:
        sessionId:
          type: string
        content:
          type: string
        guestName:
          type: string
          nullable: true
      required: [sessionId, content]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        supabaseId:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, staff, admin]
        isBanned:
          type: boolean
        createdAt:
          type: string
      required: [id, supabaseId, email, role, isBanned, createdAt]
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, staff, admin]
      required: [role]
    BanUserBody:
      type: object
      properties:
        isBanned:
          type: boolean
      required: [isBanned]
...